Projectiles use a arrayList, infinite projectile spawn, score and lives.
Found a bug that the projectile would keep going even when it was
undrawn. This caused the planet to always get hit, even when projectile
was blocked by barrier. Now projectile is in an arraylist so it can be
removed when it collides with barrier/planer.

A new projectile spawns when the old one is removed for testing
purposes/ make it slightly more like a game.

Made score and lives got down when planet hit and score go up when
projectile destroyed.